# docs/swagger.yaml
basePath: /
definitions:
...
  api.ChatCompletionMessage:
    properties:
      content:
        type: string
      role:
        type: string
    type: object
  api.ChatCompletionRequest:
    properties:
      connection_id:
        type: string
      messages:
        items:
          $ref: '#/definitions/api.ChatCompletionMessage'
        type: array
      model:
        type: string
      stream:
        type: boolean
      tools: {}
    type: object
...
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  api.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  api.LoginResponse:
    properties:
      access_token:
        type: string
    type: object
...
  api.RegisterRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  api.UserResponse:
    properties:
      id:
        description: changed from int64 to string
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Proxy server for Generative AI models.
  license:
    name: Apache 2.0
  title: GenAI Proxy API
  version: "1.0"
paths:
...
  /api/login:
    post:
      consumes:
      - application/json
      - application/x-www-form-urlencoded
      description: Log in a user with a username and password to get a JWT token.
      parameters:
      - description: User login details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - "":
        - ""
      summary: Log in a user
      tags:
      - Users
...
  /api/register:
    post:
      consumes:
      - application/json
      description: Register a new user with a username and password.
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/api.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - "":
        - ""
      summary: Register a new user
      tags:
      - Users
  /v1/chat/completions:
    post:
      consumes:
      - application/json
      description: Proxy a chat completion request to the OpenAI API.
      parameters:
      - description: OpenAI Chat Completion Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api.ChatCompletionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Proxy a chat completion request to OpenAI
      tags:
      - Proxy
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Enter your username and password to get a token.
    flow: password
    tokenUrl: /api/login
    type: oauth2
swagger: "2.0"